fix: update project for TypeScript 7.0 compatibility
TypeScript 7 removes several compiler options and changes Next.js
integration. This commit fixes all breaking changes:

- Remove `baseUrl` from both tsconfigs (removed in TS 7); update
  paths in config/tsconfig.json to use `../` prefix since they were
  relative to baseUrl: ".."
- Remove `ignoreDeprecations: "6.0"` (no longer permitted in TS 7)
- Remove `forceConsistentCasingInFileNames` (deprecated/removed)
- Add `experimental.useTypeScriptCli: true` to next.config.js so
  Next.js uses the TS CLI directly instead of its internal compiler
  API (which TS 7 removed)
- Convert jest.config.ts to jest.config.js to avoid ts-node failing
  on TS 7 (ts-node uses the removed compiler API)
- Bump typescript devDependency from ^6.0.3 to ^7.0.2

Typecheck, 214 Jest unit tests, and format all pass with TS 7.0.2.
